--- include/ogdf/graphalg/MatchingImplementation.h
@@ -1,44 +1,0 @@
-/** \file
- * \brief Reference to the current best matching implementation.
- *
- * \author Joshua Sangmeister
- *
- * \par License:
- * This file is part of the Open Graph Drawing Framework (OGDF).
- *
- * \par
- * Copyright (C)<br>
- * See README.md in the OGDF root directory for details.
- *
- * \par
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * Version 2 or 3 as published by the Free Software Foundation;
- * see the file LICENSE.txt included in the packaging of this file
- * for details.
- *
- * \par
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * \par
- * You should have received a copy of the GNU General Public
- * License along with this program; if not, see
- * http://www.gnu.org/copyleft/gpl.html
- */
-
-#pragma once
-
-#include <ogdf/graphalg/matching/MatchingBlossomV.h>
-
-namespace ogdf {
-namespace Matching {
-
-//! The current best matching algorithm implementation
-template<typename TWeight>
-using MatchingImplementation = MatchingBlossomV<TWeight>;
-
-}
-}
